requirements.txt
requests==2.23.0
terminaltables==3.1.0
python-dotenv==0.9.1
